% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_report.R
\name{daily_report}
\alias{daily_report}
\title{Get daily report data from API}
\usage{
daily_report(site, start_date, end_date, page = 1, page_size = 1000)
}
\arguments{
\item{site}{site id.}

\item{start_date}{character string for start date in ddmmyyyy format.}

\item{end_date}{character string for end date in ddmmyyyy format.}

\item{page}{page.}

\item{page_size.}{page size. Default 50.}
}
\value{
Tibble of information about traffic per day in 15 minute intervals.
}
\description{
Get daily report data from API
}
