% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_report.R
\name{monthly_report}
\alias{monthly_report}
\title{Get monthly report data from API}
\usage{
monthly_report(site, start_date, end_date, page = 1, page_size = 50)
}
\arguments{
\item{site}{site id.}

\item{start_date}{character string for start date in ddmmyyyy format.}

\item{end_date}{character string for end date in ddmmyyyy format.}

\item{page}{Page. Default 1.}

\item{page_size.}{Page size i.e. number of rows to return. Default 50.}
}
\value{
Tibble of information about traffic per day.
}
\description{
Get monthly report data from API
}
